Trabajo casi acabado. Falta contemplar copias de seguridad con su fecha y hora.

+++ BurgerKingJuanCarlos/src/burgerkingjuancarlos/ComparadorPorPrecioFinal.java
@@ -10,6 +10,6 @@ public class ComparadorPorPrecioFinal implements Comparator<Ticket> {
    
     @Override
     public int compare(Ticket p1, Ticket p2) {
-        return (p2.getPrecioTotal()- p1.getPrecioTotal());  
+        return (int) (p2.getPrecioTotal()- p1.getPrecioTotal()); 
     }
 }
